Scroll to the page's bottom on answering for smartphones
Because smartphones' displays are too small to show everything in the page,
and the contents are dynamically modified, which means the vertical
length of the page also changes everytime the user clicks answer button.
This tend to hide the messages at the bottom, which requires the user to
manually scroll down.